Bugfix - fixed memory leaks Fixed memory leak in address_pool -> not everything freed properly Fixed invalid writes in pool bitmasks -> wrong bitmask type and size of allocation caused invadlid read/writes on upper edge cases

+++ server/src/address_pool.h
@@ -17,7 +17,7 @@ typedef struct pool {
      * NOTE: Even if address is reserved, if the client is not using it it should 
      * be marked as available.
      */
-    uint32_t *leases_bm;
+    uint8_t *leases_bm;
     /* 
      * Count of curretnly not assigned addresses from this pool. 
      * Or number of 0 valued bits in leases_bm
